web_showcase/requirements.txt
Flask==3.0.0
torch==2.9.1
torchvision==0.14.1
torchaudio==2.9.1
timm==0.9.12
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
pandas==2.0.3
seaborn==0.12.2
jupyter==1.0.0
ipykernel==6.25.0
pytest==7.4.0
Werkzeug==3.0.1
